% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.contrast.R
\name{adjust.contrast}
\alias{adjust.contrast}
\title{Adjustment of contrasts}
\usage{
adjust.contrast(contrasts.matrix, contrasts.table, adjustments.list, data,
  suffix = "")
}
\arguments{
\item{contrasts.matrix}{Contrast matrix}

\item{contrasts.table}{Contrast table}

\item{adjustments.list}{Contrasts to be adjusted}

\item{data}{Data to estimate correlation of contrasts}

\item{suffix}{Suffixes added to the Group names}
}
\description{
Adjustment of contrasts
}
\examples{
adjust.contrast()
}
\keyword{adjust.contrast}
