% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustments.R
\name{adjustments}
\alias{adjustments}
\title{Expand output of GLHT, calculate FDR}
\usage{
adjustments(data)
}
\arguments{
\item{data}{}
}
\description{
Expand output of GLHT, calculate FDR
}
\examples{
adjustments(x)


}
\keyword{FDR,}
\keyword{pStars}
