% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustorder.R
\name{clustorder}
\alias{clustorder}
\title{Order factors using hierarchical clustering}
\usage{
clustorder(data, features, groups, value, dst.method = "euclidean",
  cl.method = "ward.D2", fill = 0, descending = TRUE, scalesel = "none")
}
\arguments{
\item{data}{Data which will be used in clustering}

\item{features}{Features to cluster (rows in heatmap)}

\item{groups}{Groups for values which will be clustered (columns in heatmap)}

\item{value}{Value which will be used for clustering}

\item{dst.method}{Distance method for clustering, default: 'euclidean'}

\item{cl.method}{Clustering method, default: 'ward.D2'}

\item{fill}{Value used to fill NAs}

\item{descending}{Reverse factor order?}

\item{scalesel}{Scaling ("row","col","none"), default: "none"}
}
\description{
Order factors using hierarchical clustering
}
\examples{
clustorder(data,features,groups,value,dst.method='euclidean',cl.method='ward.D2',fill=0,descending=TRUE,scalesel='none')


}
\keyword{clustorder}
