% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.adjust.R
\name{contrast.adjust}
\alias{contrast.adjust}
\title{Adjustment of contrasts}
\usage{
\method{contrast}{adjust}(contrasts.matrix, contrasts.table, adjustments.list,
  data, suffix = "")
}
\arguments{
\item{contrasts.matrix}{Contrast matrix}

\item{contrasts.table}{Contrast table}

\item{adjustments.list}{Contrasts to be adjusted}

\item{data}{Data to estimate correlation of contrasts}

\item{suffix}{Suffixes added to the Group names}
}
\description{
Adjustment of contrasts
}
\examples{
contrast.adjust()
}
\keyword{contrast.adjust}
