% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariation.R
\name{covariation}
\alias{covariation}
\title{Calculate covariation and adjustments for contrasts}
\usage{
covariation(adjustments, datam, Outliers = TRUE, Plot = FALSE)
}
\arguments{
\item{adjustments}{Adjustment table with}

\item{datam}{Contrast table}
}
\description{
Calculate covariation and adjustments for contrasts
}
\examples{
covariation()
}
\keyword{covariation}
