% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.melt.R
\name{enrichment.melt}
\alias{enrichment.melt}
\title{Melt table for enrichemnt without losing NAs}
\usage{
enrichment.melt(data, idvariables, selstats)
}
\arguments{
\item{data}{Table with data}

\item{idvariables}{Column with IDs to keep in melt}

\item{selstats}{Column with measurements to melt}
}
\description{
Melt table for enrichemnt without losing NAs
}
\examples{

}
\keyword{enrichment}
\keyword{melt}
