% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getellipse.R
\name{getellipse}
\alias{getellipse}
\title{Calculate ellipse for 2D distirbution}
\usage{
getellipse(x, y, sc = 1)
}
\arguments{
\item{x}{Values in 1st dimension}

\item{y}{Values in 1st dimension}

\item{sc}{Scaling factor to SD}
}
\description{
Calculate ellipse for 2D distirbution
}
\examples{

}
\keyword{Ellipse}
