% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.filter.R
\name{group.filter}
\alias{group.filter}
\title{Adjustment of contrasts}
\usage{
group.filter(dt, thrs)
}
\arguments{
\item{dt}{edgeR DGEList object}

\item{thrs}{CPM theshold for groups}
}
\description{
Adjustment of contrasts
}
\examples{
group.filter()
}
\keyword{adjust.contrast}
