% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMap.R
\name{HMap}
\alias{HMap}
\title{Draw heatmap}
\usage{
HMap(data, ID, feature, value, info, cols = list(), scalesel = "row",
  rnames = FALSE, reorder.rows = TRUE, reorder.columns = TRUE,
  name = "Z-score", cdr = "euclidean", cdc = "euclidean",
  cmr = "ward.D2", cmc = "ward.D2", rmax = 10)
}
\arguments{
\item{data}{Data used for heatmap}

\item{ID}{Observations in columns}

\item{feature}{feature in rows}

\item{value}{Value to be used in heatmap}

\item{info}{Table with descriptors of interest}

\item{cols}{Colorcoding list}
}
\description{
Draw heatmap
}
\examples{
HMap(data,ID,feature,value,info,cols=list(),scalesel='row',rnames=FALSE,reorder.rows=TRUE,reorder.columns=TRUE,name='Z-score',title="Heatmap",cdr="euclidean",cdc="euclidean",cmr="ward.D2",cmc="ward.D2",rmax=10)
}
\keyword{clustorder}
