% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesise.R
\name{hypothesise}
\alias{hypothesise}
\title{Linear Hypothesis Testing in HT manner using Dplyr}
\usage{
hypothesise(lmdata, formula, cont.matrix, weights.col = NA, variable = NA,
  verbose = FALSE)
}
\arguments{
\item{formula}{Formula for linear model in the form of "Values~0+Group"}

\item{cont.matrix}{Contrast matrix with dummy variables.}

\item{weights.col}{Column with weights for observations. Default: NA}

\item{variable}{Name of the column with variables to print. Default: NA}

\item{verbose}{Print dataset completeness}

\item{lmshape}{Table containing samples in rows and variables in columns}
}
\description{
Linear Hypothesis Testing in HT manner using Dplyr
}
\examples{
hypothesise2()
}
\keyword{hypothesise}
