% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesise2.R
\name{hypothesise2}
\alias{hypothesise2}
\title{Linear Hypothesis Testing in HT manner using Dplyr}
\usage{
hypothesise2(lmdata, formula, cont.matrix, weights.col = NA, variable = NA)
}
\arguments{
\item{formula}{Formula for linear model in the form of "Values~0+Group"}

\item{cont.matrix}{Contrast matrix with dummy variables.}

\item{weights.col}{Column with weights for observations. Default: NA}

\item{lmshape}{Table containing samples in rows and variables in columns}

\item{variables}{Name of the column with variables to print. Default: NA}
}
\description{
Linear Hypothesis Testing in HT manner using Dplyr
}
\examples{
hypothesise2()
}
\keyword{hypothesise}
