% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmtest.R
\name{lmtest}
\alias{lmtest}
\title{Simple linear modelling with 2-way interaction}
\usage{
lmtest(data, formula)
}
\arguments{
\item{data}{Table containing samples in rows and variables in columns}

\item{formula}{Formula for linear model in the form of "Values~Factor1*Factor2"}
}
\description{
Simple linear modelling with 2-way interaction
}
\examples{
lmtest
}
\keyword{hypothesise}
