% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makereplicates.R
\name{makereplicates}
\alias{makereplicates}
\title{Make a vector with unique replicate indexes}
\usage{
makereplicates(data)
}
\arguments{
\item{data}{Table containing columns over which unique repicate indices need to be generated}
}
\description{
Make a vector with unique replicate indexes
}
\examples{
hypothesise()
}
\keyword{makereplicates}
