% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinMeanSDMax.R
\name{MinMeanSDMax}
\alias{MinMeanSDMax}
\title{Calculates Mean, SD, Min, Max for boxplot}
\usage{
MinMeanSDMax(x)
}
\arguments{
\item{x}{Vector with numeric values}
}
\description{
Calculates Mean, SD, Min, Max for boxplot
}
\examples{

}
\keyword{Max}
\keyword{Mean,}
\keyword{Min,}
\keyword{SD,}
