% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplex.R
\name{multiplex}
\alias{multiplex}
\title{Multiplex your data frame}
\usage{
multiplex(data, groups, dims = 3)
}
\arguments{
\item{data}{Data to multiplex}

\item{groups}{Grouping variables to retain, everything else will be multiplexed}

\item{dim}{Number of dimensions in the final table. Minimum is two, anything more than 4 can take a very long time...}
}
\description{
Multiplex your data frame
}
\examples{
multiplex(data,groups,dim=)


}
\keyword{multiplex}
