% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAprep.R
\name{PCAprep}
\alias{PCAprep}
\title{Draw heatmap}
\usage{
PCAprep(data, ID, feature, value, info, grouping = "All", scaling = TRUE,
  dmethod = "euclidean", cmethod = "ward.D2")
}
\arguments{
\item{data}{Data used for heatmap}

\item{ID}{Observations in columns}

\item{feature}{feature in rows}

\item{value}{Value to be used in heatmap}

\item{info}{Table with descriptors of interest}

\item{grouping}{Groups of observations for ellipses}

\item{dmethod}{Distance measure for hierarchical clustering}

\item{cmethod}{Clustering method for hierarchical clustering}

\item{scalink}{Scale feature for hierarchical clustering?}
}
\description{
Draw heatmap
}
\examples{
PCAprep(data,ID,feature,value,info,cols=list(),scalesel='row',rnames=FALSE,reorder.rows=TRUE,reorder.columns=TRUE,name='Z-score',title="Heatmap",cdr="euclidean",cdc="euclidean",cmr="ward.D2",cmc="ward.D2",rmax=10)
}
\keyword{PCAprep}
