% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotEnrichment.R
\name{PlotEnrichment}
\alias{PlotEnrichment}
\title{Plot Enrichment data}
\usage{
PlotEnrichment(data, xval, yval, fillval = "logFDRbin", ncols = 6)
}
\arguments{
\item{data}{Data for plotting}

\item{xval}{Comparisons}

\item{yval}{Features}

\item{fillval}{Value for fill, factor}

\item{ncols}{Number of color levels, default: 5. length(enrbrks)}
}
\description{
Plot Enrichment data
}
\examples{
PlotEnrichment(data,xval,yval,fillval,ncols=5)

}
\keyword{PlotEnrichment}
