% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pStars.R
\name{pStars}
\alias{pStars}
\title{Get clean significance stars}
\usage{
pStars(x)
}
\arguments{
\item{x}{Vector witg p values}
}
\description{
Get clean significance stars
}
\examples{
pStars(x)


}
\keyword{Stars}
\keyword{p,}
