% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.and.clean.R
\name{read.and.clean}
\alias{read.and.clean}
\title{Read Excel and csv files}
\usage{
read.and.clean(fname, sheet = "", sep = ",")
}
\arguments{
\item{fname}{Table containing sampels in rows and variables in columns}

\item{sheet}{Sheet to be used}

\item{sep}{Separator in csv file. DEFAULTS to ","}
}
\description{
Read Excel and csv files
}
\examples{
read.and.clean()
}
\keyword{read.and.clean}
