% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.contrasts2.R
\name{read.contrasts2}
\alias{read.contrasts2}
\title{Get contrasts for Linear Hypothesis Testing in HT manner.
Works with hypothesise2}
\usage{
read.contrasts2(cfile, csheet)
}
\arguments{
\item{cfile}{Excel file containing standard contrast table}

\item{csheet}{Sheet in Excel file containing contrast table}

\item{samples.selected}{Limit contrasts to selected samples}

\item{variables}{Variables to preserve in the table}
}
\description{
Get contrasts for Linear Hypothesis Testing in HT manner.
Works with hypothesise2
}
\examples{
read.contrasts()


}
\keyword{contrasts}
\keyword{read.contrasts}
