% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_PN.R
\name{theme_PN}
\alias{theme_PN}
\title{Theme Publication}
\usage{
theme_PN(base_size = par()$ps, base_family = par()$family)
}
\arguments{
\item{base_size}{Base size}

\item{base_family}{Base family}
}
\description{
Theme Publication
}
\examples{

}
\keyword{Publication}
\keyword{ggtheme,}
