% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_Publication.R
\name{theme_Publication}
\alias{theme_Publication}
\title{Theme Publication}
\usage{
theme_Publication(base_size = 14)
}
\arguments{
\item{base_size}{Base size}
}
\description{
Theme Publication
}
\examples{

}
\keyword{Publication}
\keyword{ggtheme,}
