% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.adjust.R
\name{contrast.adjust}
\alias{contrast.adjust}
\title{Calculate covariation and adjustments for contrasts}
\usage{
contrast.adjust(contr.mat, adjustments)
}
\arguments{
\item{contr.mat}{Contrast matrix with contrast names as rownames}

\item{adjustments}{Adjustment table generated with "covariation" function}
}
\description{
Calculate covariation and adjustments for contrasts
}
\examples{
contrast.adjust()
}
\keyword{adjust,}
\keyword{contrast}
\keyword{contrast.adjust,}
