% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{enrichment}
\alias{enrichment}
\title{Calculate enrichment}
\usage{
enrichment(data, grouping, feature, Sbrks = c(0, -log(0.05, 10), 2, 3, 4,
  1000), Slbls = c("N.S.", "<0.05", "<0.01", "<0.001", "<0.0001"))
}
\arguments{
\item{data}{Table with data}

\item{grouping}{Columns over which data needs to be grouped}

\item{feature}{Feature for which enrichment is to be calculated}

\item{Sbrks}{Enrichment breaks in -log10 scale}

\item{Slbls}{Enrichment labels (one less than breaks)}
}
\description{
Calculate enrichment
}
\examples{

}
\keyword{enrichment}
