% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getresults.R
\name{getresults}
\alias{getresults}
\title{Get formated results from LM dataset}
\usage{
getresults(data, contrasts.desc)
}
\arguments{
\item{data}{LM output (do not ungroup it!)}

\item{contrasts.desc}{Contrast descriptors}
}
\description{
Get formated results from LM dataset
}
\examples{
getresults(data,contrasts.desc)


}
\keyword{results}
