% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesise.R
\name{hypothesise}
\alias{hypothesise}
\title{Linear Hypothesis Testing in HT manner}
\usage{
hypothesise(lmshape, variables, cont.matrix, fml = "0+Group",
  weights.mat = NA)
}
\arguments{
\item{lmshape}{Table containing samples in rows and variables in columns}

\item{variables}{Variables to be tested in model}

\item{cont.matrix}{Contrast matrix with dummy variables}

\item{fml}{Formula that will be used to spread the effects of samples. DEFAULT to "0+Group"}
}
\description{
Linear Hypothesis Testing in HT manner
}
\examples{
hypothesise()
}
\keyword{hypothesise}
