% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesise2.R
\name{hypothesise2}
\alias{hypothesise2}
\title{Linear Hypothesis Testing in HT manner using Dplyr}
\usage{
hypothesise2(lmdata, formula, cont.matrix, weights.col = NA, variable = NA)
}
\arguments{
\item{cont.matrix}{Contrast matrix with dummy variables}

\item{lmshape}{Table containing samples in rows and variables in columns}

\item{variables}{Variables to be tested in model}

\item{fml}{Formula that will be used to spread the effects of samples. DEFAULT to "0+Group"}
}
\description{
Linear Hypothesis Testing in HT manner using Dplyr
}
\examples{
hypothesise2()
}
\keyword{hypothesise}
