% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.contrasts.R
\name{read.contrasts}
\alias{read.contrasts}
\title{Linear Hypothesis Testing in HT manner}
\usage{
read.contrasts(cfile, csheet, samples.selected, variables = c())
}
\arguments{
\item{cfile}{Excel file containing standard contrast table}

\item{csheet}{Sheet in Excel file containing contrast table}

\item{samples.selected}{Limit contrasts to selected samples}

\item{variables}{Other variables to preserve in table}
}
\description{
Linear Hypothesis Testing in HT manner
}
\examples{
read.contrasts()


}
\keyword{contrasts}
\keyword{read.contrasts}
